Moving common structs and enums to ShaderTestCommon.h
